Improvements in order support cabal v2-build (fixes #26).
This belatedly adds support for building Qtah via Cabal's v2-* build
commands.  Several changes were necessary to get this to work.

The way Hoppy generators work, as executables, doesn't play very well with
Cabal.  In order for -cpp and -hs packages to be able to call their
generator properly, they would need to be able to declare a dependency on
the generator's executable.  Cabal doesn't truly support this though.
There's a proposal for a setup-tool-depends directive[1], with a
workaround, but what's easier to do is to link to the code needed as a
library via setup-depends.

So the first change is to make the generator callable entirely as a
library.  This is done through new 'generateCpp' and 'generateHs' functions
in Graphics.UI.Qtah.Generator.Main.

This also requires porting the qtah-listener-gen shell script to Haskell.
And thanks to this, we don't need to install qtah-listener-gen specially
any more as part of qtah-generator, so qtah-generator can go back to using
the Simple build type.

Additionally, this adds a missing register hook to qtah-cpp, as needed for
the in-place registrations that v2-build does.  This was fixed in Hoppy two
years ago but didn't get ported over to Qtah.

qtah-generator is still buildable and runnable as an executable, but this
is no longer used by the build system.

[1] https://github.com/haskell/cabal/issues/4648